Add x/demo package.
This is a working demonstration harness for the feed/hub prototype. Running the
program starts a JSON-RPC service (over HTTP) that allows the caller to post
events, query the index, and subscribe and consume items from the hub via
listeners.

Usage:
   demo -listen [host]:port

The service endpoint is http://[host]:port/rpc.
Content-Type: application/json
Content: JSON-RPC requests or batches.

To simulate an active node, the server runs a "heartbeat" that periodically
injects new items into the feed.